Added mps support to fine tune easy ocr and fixed deprecated torch function _accumulate. Also working on a requirement.txt for training.
