+++ requirements.txt
@@ -10,3 +10,6 @@ PyYAML
 Shapely
 pyclipper
 ninja
+six
+tqdm
+editdistance
