+++ saved_models/en_filtered/log_train.txt
@@ -1,0 +1,90 @@
+[2/10] Train loss: 71.21156, Valid loss: 13.95401, Elapsed_time: 2.35558
+Current_accuracy : 0.000, Current_norm_ED  : 0.0000
+Best_accuracy    : 0.000, Best_norm_ED     : 0.0000
+--------------------------------------------------------------------------------
+Ground Truth              | Prediction                | Confidence Score & T/F
+--------------------------------------------------------------------------------
+28.58                     |                           | 0.0000	False
+CASH                      |                           | 0.0000	False
+--------------------------------------------------------------------------------
+[4/10] Train loss: 5.49628, Valid loss: 67.17086, Elapsed_time: 463.18695
+Current_accuracy : 0.516, Current_norm_ED  : 0.0745
+Best_accuracy    : 0.516, Best_norm_ED     : 0.0745
+--------------------------------------------------------------------------------
+Ground Truth              | Prediction                | Confidence Score & T/F
+--------------------------------------------------------------------------------
+11.00 S                   |                           | 0.0000	False
+8.70                      | 0                         | 0.0000	False
+--------------------------------------------------------------------------------
+[6/10] Train loss: 21.82625, Valid loss: 5.25841, Elapsed_time: 1005.17056
+Current_accuracy : 0.000, Current_norm_ED  : 0.0000
+Best_accuracy    : 0.516, Best_norm_ED     : 0.0745
+--------------------------------------------------------------------------------
+Ground Truth              | Prediction                | Confidence Score & T/F
+--------------------------------------------------------------------------------
+6.00 Z                    |                           | 0.1271	False
+AMOUNT                    |                           | 0.0833	False
+--------------------------------------------------------------------------------
+[8/10] Train loss: 6.66635, Valid loss: 5.06618, Elapsed_time: 1466.72036
+Current_accuracy : 0.000, Current_norm_ED  : 0.0000
+Best_accuracy    : 0.516, Best_norm_ED     : 0.0745
+--------------------------------------------------------------------------------
+Ground Truth              | Prediction                | Confidence Score & T/F
+--------------------------------------------------------------------------------
+: CS00010855              |                           | 0.1149	False
+TOTAL :                   |                           | 0.1542	False
+--------------------------------------------------------------------------------
+[10/10] Train loss: 6.81450, Valid loss: 4.88509, Elapsed_time: 1926.84612
+Current_accuracy : 0.000, Current_norm_ED  : 0.0000
+Best_accuracy    : 0.516, Best_norm_ED     : 0.0745
+--------------------------------------------------------------------------------
+Ground Truth              | Prediction                | Confidence Score & T/F
+--------------------------------------------------------------------------------
+VE05: RIDZUAN (11900)     |                           | 0.0464	False
+AMOUNT                    |                           | 0.1669	False
+--------------------------------------------------------------------------------
+[10/10] Train loss: 55.63767, Valid loss: 20.73927, Elapsed_time: 1.58839
+Current_accuracy : 0.000, Current_norm_ED  : 0.0000
+Best_accuracy    : 0.000, Best_norm_ED     : 0.0000
+--------------------------------------------------------------------------------
+Ground Truth              | Prediction                | Confidence Score & T/F
+--------------------------------------------------------------------------------
+3                         |                           | 0.0000	False
+GST SUMMARY               |                           | 0.0000	False
+--------------------------------------------------------------------------------
+[1/2] Train loss: 3.35850, Valid loss: 3.24225, Elapsed_time: 1.89526
+Current_accuracy : 35.967, Current_norm_ED  : 0.7093
+Best_accuracy    : 35.967, Best_norm_ED     : 0.7093
+--------------------------------------------------------------------------------
+Ground Truth              | Prediction                | Confidence Score & T/F
+--------------------------------------------------------------------------------
+002043319-W               | 002043*19-W               | 0.0710	False
+TOTAL PAYABLE:            | Total Payable:            | 0.3802	False
+--------------------------------------------------------------------------------
+[2/2] Train loss: 2.43363, Valid loss: 2.97656, Elapsed_time: 510.12301
+Current_accuracy : 35.972, Current_norm_ED  : 0.7054
+Best_accuracy    : 35.972, Best_norm_ED     : 0.7093
+--------------------------------------------------------------------------------
+Ground Truth              | Prediction                | Confidence Score & T/F
+--------------------------------------------------------------------------------
+432.60                    | 432.60                    | 0.8497	True
+GST SUMMARY               | GST Summary               | 0.4686	False
+--------------------------------------------------------------------------------
+[1/2] Train loss: 3.09099, Valid loss: 3.27444, Elapsed_time: 15.87064
+Current_accuracy : 35.751, Current_norm_ED  : 0.7089
+Best_accuracy    : 35.751, Best_norm_ED     : 0.7089
+--------------------------------------------------------------------------------
+Ground Truth              | Prediction                | Confidence Score & T/F
+--------------------------------------------------------------------------------
+19/05/2018 09:04:55 PM    | 19/026/2018 09.04.55 pm   | 0.0433	False
+THANK YOU FOR SHOPPING AT PASARAYA | Thank vou For shopping ak PASARAVA | 0.1214	False
+--------------------------------------------------------------------------------
+[2/2] Train loss: 2.91665, Valid loss: 2.99240, Elapsed_time: 509.35107
+Current_accuracy : 35.656, Current_norm_ED  : 0.7062
+Best_accuracy    : 35.751, Best_norm_ED     : 0.7089
+--------------------------------------------------------------------------------
+Ground Truth              | Prediction                | Confidence Score & T/F
+--------------------------------------------------------------------------------
+20.20                     | 20.20                     | 0.7442	True
+5.60                      | 5.60                      | 0.9916	True
+--------------------------------------------------------------------------------
