--- trainer/all_data/folder.txt
@@ -1,1 +1,0 @@
-place dataset folder here
