+++ trainer/config_files/en_filtered_config.yaml
@@ -2,14 +2,14 @@ number: '0123456789'
 symbol: "!\"#$%&'()*+,-./:;<=>?@[\\]^_`{|}~ €"
 lang_char: 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz'
 experiment_name: 'en_filtered'
-train_data: 'all_data'
-valid_data: 'all_data/en_val'
+train_data: 'trainer/all_data/train'
+valid_data: 'trainer/all_data/test'
 manualSeed: 1111
-workers: 6
-batch_size: 32 #32
-num_iter: 300000
-valInterval: 20000
-saved_model: '' #'saved_models/en_filtered/iter_300000.pth'
+workers: 13
+batch_size: 256 #32
+num_iter: 2
+valInterval: 2
+saved_model: 'trainer/saved_models/english_g2.pth'
 FT: False
 optim: False # default is Adadelta
 lr: 1.
@@ -17,8 +17,9 @@ beta1: 0.9
 rho: 0.95
 eps: 0.00000001
 grad_clip: 5
+
 #Data processing
-select_data: 'en_train_filtered' # this is dataset folder in train_data
+select_data: '' # this is dataset folder in train_data
 batch_ratio: '1' 
 total_data_usage_ratio: 1.0
 batch_max_length: 34 
@@ -30,6 +31,7 @@ sensitive: True
 PAD: True
 contrast_adjust: 0.0
 data_filtering_off: False
+
 # Model Architecture
 Transformation: 'None'
 FeatureExtraction: 'VGG'
@@ -42,4 +44,4 @@ hidden_size: 256
 decode: 'greedy'
 new_prediction: False
 freeze_FeatureFxtraction: False
-freeze_SequenceModeling: False
+freeze_SequenceModeling: True
