+++ trainer/dataset.py
@@ -10,7 +10,7 @@ from natsort import natsorted
 from PIL import Image
 import numpy as np
 from torch.utils.data import Dataset, ConcatDataset, Subset
-from torch._utils import _accumulate
+from itertools import accumulate as _accumulate
 import torchvision.transforms as transforms
 
 def contrast_grey(img):
@@ -78,7 +78,7 @@ class Batch_Balanced_Dataset(object):
                 _dataset, batch_size=_batch_size,
                 shuffle=True,
                 num_workers=int(opt.workers), #prefetch_factor=2,persistent_workers=True,
-                collate_fn=_AlignCollate, pin_memory=True)
+                collate_fn=_AlignCollate, pin_memory=False)
             self.data_loader_list.append(_data_loader)
             self.dataloader_iter_list.append(iter(_data_loader))
 
@@ -98,12 +98,12 @@ class Batch_Balanced_Dataset(object):
 
         for i, data_loader_iter in enumerate(self.dataloader_iter_list):
             try:
-                image, text = data_loader_iter.next()
+                image, text = next(data_loader_iter)
                 balanced_batch_images.append(image)
                 balanced_batch_texts += text
             except StopIteration:
                 self.dataloader_iter_list[i] = iter(self.data_loader_list[i])
-                image, text = self.dataloader_iter_list[i].next()
+                image, text = next(self.dataloader_iter_list[i])
                 balanced_batch_images.append(image)
                 balanced_batch_texts += text
             except ValueError:
