--- trainer/saved_models/folder.txt
@@ -1,1 +1,0 @@
-trained model will be saved here
